fix unit test problem with a clean separation
